:recycle: (deeplink) respect architecture std for generateLongFirebaseDynamicLink to be moved to helpers
