+++ src/features/deeplinks/index.ts
@@ -1,1 +1,1 @@
-export * from './utils'
+export * from './generateLongFirebaseDynamicLink'
